.github: Drop sys-boot/gnu-efi from automation for now
Updating it to 3.0.17 breaks the build. First there are some
signedness issues (which I fixed), but then there is some header
ordering issue which results in errors about some names being
undefined. This is because a header that defines those names includes
some other header before defining the names. That other header pulls
in some other header that end up referencing the not-yet-defined
names:

In file included from /usr/include/efi/efilib.h:22,
                 from Include/OpenSslSupport.h:19,
                 from Include/stddef.h:15,
                 from /usr/include/efi/x86_64/efibind.h:99,
                 from /usr/include/efi/efi.h:42,
                 from ./Library/BaseLib.h:1,
                 from ./InternalCryptLib.h:18,
                 from Hash/CryptMd4.c:15:
/usr/include/efi/efidebug.h:22:8: error: unknown type name ‘UINTN’
   22 | extern UINTN     EFIDebug;

/usr/include/efi/x86_64/efibind.h defines UINTN in line 124, but
includes stddef.h in line 99.

I'm not dealing with it knowing that Sayan is working on secure-boot
stuff which involves updating sys-boot/shim, so this bug should go
away soon.