overlay profiles: Update accept keywords for dev-libs/libdnet
There was a change in libdnet ebuilds about how tests are
enabled/disabled through a configure flag. The change was correct for
versions 1.16.4 and up, but not for earlier ones. So instead of
waiting for the change to be reverted for older ebuilds, bump the
version to 1.16.4 now - it should be stabilized soon.